logging package + all updates in libs
