+++ libs/adapters/src/index.ts
@@ -1,7 +1,6 @@
 export * as express from './express';
 export * from './hot-shots';
 export * from './mixpanel';
-export * from './pino';
 export * from './rabbitmq';
 export * from './redis';
 export * as trpc from './trpc';
