+++ libs/adapters/src/rabbitmq/RabbitMQAdapter.ts
@@ -4,13 +4,12 @@ import {
   EventContext,
   EventSchemas,
   EventsHandlerMetadata,
-  ILogger,
   InvalidInput,
   RetryStrategyFn,
   eventHandler,
-  logger,
   schemas,
 } from '@hicommonwealth/core';
+import { ILogger, logger } from '@hicommonwealth/logging';
 import { Message } from 'amqplib';
 import * as Rascal from 'rascal';
 import { AckOrNack } from 'rascal';
@@ -61,7 +60,7 @@ export class RabbitMQAdapter implements Broker {
   private _log: ILogger;
 
   constructor(protected readonly _rabbitMQConfig: Rascal.BrokerConfig) {
-    this._log = logger().getLogger(__filename);
+    this._log = logger(__filename);
     this._rawVhost =
       _rabbitMQConfig.vhosts![Object.keys(_rabbitMQConfig.vhosts!)[0]];
     this.subscribers = Object.keys(this._rawVhost.subscriptions);
