+++ libs/adapters/src/redis/RedisCacheAdapter.ts
@@ -1,10 +1,5 @@
-import {
-  Cache,
-  ILogger,
-  delay,
-  logger,
-  type CacheNamespaces,
-} from '@hicommonwealth/core';
+import { Cache, delay, type CacheNamespaces } from '@hicommonwealth/core';
+import { ILogger, logger } from '@hicommonwealth/logging';
 import { RedisClientOptions, createClient, type RedisClientType } from 'redis';
 
 const CONNECT_TIMEOUT = 5000;
@@ -48,7 +43,7 @@ export class RedisCache implements Cache {
       };
     }
 
-    this._log = logger().getLogger(__filename);
+    this._log = logger(__filename);
     this._log.info(`Connecting to Redis at: ${redis_url}`);
     this._client = createClient(redisOptions) as RedisClientType;
