+++ libs/adapters/src/redis/RedisCacheDecorator.ts
@@ -1,4 +1,5 @@
-import { CacheNamespaces, ILogger, cache, logger } from '@hicommonwealth/core';
+import { CacheNamespaces, cache } from '@hicommonwealth/core';
+import { ILogger, logger } from '@hicommonwealth/logging';
 import { NextFunction, Request, RequestHandler, Response } from 'express';
 import {
   CacheKeyDuration,
@@ -33,7 +34,7 @@ export class CacheDecorator {
   private _disabled = false;
 
   constructor() {
-    this._log = logger().getLogger(__filename);
+    this._log = logger(__filename);
     // If cache is disabled, skip caching
     if (process.env.DISABLE_CACHE === 'true') {
       this._log.info(`cacheMiddleware: cache disabled`);
