+++ libs/adapters/tsconfig.json
@@ -7,5 +7,5 @@
     "strict": true
   },
   "include": ["src"],
-  "references": [{ "path": "../core" }]
+  "references": [{ "path": "../core" }, { "path": "../logging" }]
 }
