+++ libs/core/src/ports/interfaces.ts
@@ -1,3 +1,4 @@
+import { ILogger } from '@hicommonwealth/logging';
 import {
   EventContext,
   EventSchemas,
@@ -23,33 +24,6 @@ export interface Disposable {
  * Adapter factory
  */
 export type AdapterFactory<T extends Disposable> = (adapter?: T) => T;
-
-export type LogContext = {
-  // fingerprint is a Rollbar concept that helps Rollbar group error occurrences together
-  fingerprint?: string;
-  [key: string]: unknown;
-};
-
-/**
- * Logger port
- * Logs messages at different levels
- */
-export interface ILogger {
-  trace(msg: string, error?: Error, context?: LogContext): void;
-  debug(msg: string, error?: Error, context?: LogContext): void;
-  info(msg: string, error?: Error, context?: LogContext): void;
-  warn(msg: string, error?: Error, context?: LogContext): void;
-  error(msg: string, error?: Error, context?: LogContext): void;
-  fatal(msg: string, error?: Error, context?: LogContext): void;
-}
-
-/**
- * Logger factory
- * Builds a named logger
- */
-export interface Logger extends Disposable {
-  getLogger(...ids: string[]): ILogger;
-}
 
 /**
  * Stats port
