+++ libs/core/src/ports/port.ts
@@ -1,6 +1,6 @@
+import { logger } from '@hicommonwealth/logging';
 import { ExitCode } from './enums';
 import { successfulInMemoryBroker } from './in-memory-brokers';
-import { getInMemoryLogger } from './in-memory-logger';
 import {
   AdapterFactory,
   Analytics,
@@ -8,9 +8,10 @@ import {
   Cache,
   Disposable,
   Disposer,
-  Logger,
   Stats,
 } from './interfaces';
+
+const log = logger(__filename);
 
 /**
  * Map of disposable adapter instances
@@ -27,9 +28,7 @@ export function port<T extends Disposable>(factory: AdapterFactory<T>) {
     if (!adapters.has(factory.name)) {
       const instance = factory(adapter);
       adapters.set(factory.name, instance);
-      logger()
-        .getLogger(__filename)
-        .info(`[binding adapter] ${instance.name || factory.name}`);
+      log.info(`[binding adapter] ${instance.name || factory.name}`);
       return instance;
     }
     return adapters.get(factory.name) as T;
@@ -53,9 +52,7 @@ const disposeAndExit = async (code: ExitCode = 'UNIT_TEST'): Promise<void> => {
   await Promise.all(disposers.map((disposer) => disposer()));
   await Promise.all(
     [...adapters].reverse().map(async ([key, adapter]) => {
-      logger()
-        .getLogger(__filename)
-        .info(`[disposing adapter] ${adapter.name || key}`);
+      log.info(`[disposing adapter] ${adapter.name || key}`);
       await adapter.dispose();
     }),
   );
@@ -81,31 +78,20 @@ export const dispose = (
  * Handlers to dispose registered resources on exit or unhandled exceptions
  */
 process.once('SIGINT', async (arg?: any) => {
-  logger()
-    .getLogger(__filename)
-    .info(`SIGINT ${arg !== 'SIGINT' ? arg : ''}`);
+  log.info(`SIGINT ${arg !== 'SIGINT' ? arg : ''}`);
   await disposeAndExit('EXIT');
 });
 process.once('SIGTERM', async (arg?: any) => {
-  logger()
-    .getLogger(__filename)
-    .info(`SIGTERM ${arg !== 'SIGTERM' ? arg : ''}`);
+  log.info(`SIGTERM ${arg !== 'SIGTERM' ? arg : ''}`);
   await disposeAndExit('EXIT');
 });
 process.once('uncaughtException', async (arg?: any) => {
-  logger().getLogger(__filename).error('Uncaught Exception', arg);
+  log.error('Uncaught Exception', arg);
   await disposeAndExit('ERROR');
 });
 process.once('unhandledRejection', async (arg?: any) => {
-  logger().getLogger(__filename).error('Unhandled Rejection', arg);
+  log.error('Unhandled Rejection', arg);
   await disposeAndExit('ERROR');
-});
-
-/**
- * Logger port factory
- */
-export const logger = port(function logger(logger?: Logger) {
-  return logger || getInMemoryLogger();
 });
 
 /**
