+++ libs/logging/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@hicommonwealth/logging",
+  "private": "true",
+  "version": "0.1.0",
+  "description": "Logging Utilities",
+  "files": [
+    "build"
+  ],
+  "main": "build/index",
+  "types": "src/index",
+  "scripts": {
+    "build": "tsc -b",
+    "clean": "rm -rf build",
+    "check-types": "tsc --noEmit"
+  },
+  "dependencies": {
+    "pino": "^8.19.0",
+    "pino-http": "^9.0.0",
+    "pino-http-print": "^3.1.0"
+  },
+  "devDependencies": {
+    "pino-pretty": "^10.3.1"
+  }
+}
