+++ libs/model/src/models/comment.ts
@@ -1,4 +1,5 @@
-import { IDiscordMeta, logger, stats } from '@hicommonwealth/core';
+import { IDiscordMeta, stats } from '@hicommonwealth/core';
+import { logger } from '@hicommonwealth/logging';
 import type * as Sequelize from 'sequelize';
 import type { DataTypes } from 'sequelize';
 import type { AddressAttributes } from './address';
@@ -6,7 +7,7 @@ import type { CommunityAttributes } from './community';
 import { ThreadAttributes } from './thread';
 import type { ModelInstance, ModelStatic } from './types';
 
-const log = logger().getLogger(__filename);
+const log = logger(__filename);
 
 export type CommentAttributes = {
   thread_id: string;
