+++ libs/model/src/services/tokenBalanceCache/providers/getEvmBalances.ts
@@ -1,4 +1,5 @@
-import { BalanceSourceType, logger } from '@hicommonwealth/core';
+import { BalanceSourceType } from '@hicommonwealth/core';
+import { logger } from '@hicommonwealth/logging';
 import Web3 from 'web3';
 import { models } from '../../../database';
 import { Balances, GetEvmBalancesOptions } from '../types';
@@ -8,7 +9,7 @@ import { __getErc20Balances } from './get_erc20_balances';
 import { __getErc721Balances } from './get_erc721_balances';
 import { __getEthBalances } from './get_eth_balances';
 
-const log = logger().getLogger(__filename);
+const log = logger(__filename);
 
 export async function getEvmBalances(
   options: GetEvmBalancesOptions,
