+++ libs/model/src/services/tokenBalanceCache/providers/get_cosmos_balances.ts
@@ -6,12 +6,12 @@ import {
   setupBankExtension,
   setupStakingExtension,
 } from '@cosmjs/stargate';
-import { logger } from '@hicommonwealth/core';
+import { logger } from '@hicommonwealth/logging';
 import { ChainNodeInstance } from '../../../models/chain_node';
 import { Balances } from '../types';
 import { getTendermintClient } from '../util';
 
-const log = logger().getLogger(__filename);
+const log = logger(__filename);
 
 export type GetCosmosNativeBalanceOptions = {
   chainNode: ChainNodeInstance;
