+++ libs/model/src/services/tokenBalanceCache/providers/get_cw20_balances.ts
@@ -1,11 +1,12 @@
 import { WasmExtension, setupWasmExtension } from '@cosmjs/cosmwasm-stargate';
 import { QueryClient } from '@cosmjs/stargate';
-import { logger } from '@hicommonwealth/core';
 import { ChainNodeInstance } from '../../../models/chain_node';
 import { Balances } from '../types';
 import { getTendermintClient } from '../util';
 
-const log = logger().getLogger(__filename);
+import { logger } from '@hicommonwealth/logging';
+
+const log = logger(__filename);
 
 export type GetCw20BalancesOptions = {
   chainNode: ChainNodeInstance;
