+++ libs/model/src/services/tokenBalanceCache/providers/get_erc1155_balances.ts
@@ -1,10 +1,10 @@
-import { logger } from '@hicommonwealth/core';
+import { logger } from '@hicommonwealth/logging';
 import AbiCoder from 'web3-eth-abi';
 import { ChainNodeInstance } from '../../../models/chain_node';
 import { Balances } from '../types';
 import { evmRpcRequest } from '../util';
 
-const log = logger().getLogger(__filename);
+const log = logger(__filename);
 
 export type GetErc1155BalancesOptions = {
   chainNode: ChainNodeInstance;
